Solutions/2024/AdventOfCode2024.Tests/Day22/Day22SolverTests.cs
namespace AdventOfCode2024.Day22.Tests;

public class Day22SolverTests
{
    public static TheoryData<int, string> PuzzleOneTestData
        = new TheoryData<int, string>()
        {
                {
                    37327623,
                    """
                    1
                    10
                    100
                    2024
                    """
                }
        };

    [Theory]
    [MemberData(nameof(PuzzleOneTestData))]
    public void PuzzleOne(int expected, string input)
    {
        var actual = Day22Solver.PuzzleOne(input);

        Assert.Equal(expected, actual);
    }

    public static TheoryData<int, string> PuzzleTwoTestData
        = new TheoryData<int, string>()
        {
                {
                    23,
                    """
                    1
                    2
                    3
                    2024
                    """
                }
        };

    [Theory]
    [MemberData(nameof(PuzzleTwoTestData))]
    public void PuzzleTwo(int expected, string input)
    {
        var actual = Day22Solver.PuzzleTwo(input);

        Assert.Equal(expected, actual);
    }
}